Separate the tests file for diffgeom in multiple files.
