+++ sympy/diffgeom/tests/test_function_diffgeom_book.py
@@ -106,27 +106,3 @@ def test_functional_diffgeom_ch4():
     assert simplify(R2.dtheta(circ)(p_r)) == 1
 
     assert (circ - R2.e_theta)(s_field_r)(p_r) == 0
-
-
-def test_R2():
-    x0, y0, r0, theta0 = symbols('x0, y0, r0, theta0', real=True)
-    point_r = R2_r.point([x0, y0])
-    point_p = R2_p.point([r0, theta0])
-
-    # r**2 = x**2 + y**2
-    assert (R2.r**2 - R2.x**2 - R2.y**2)(point_r) == 0
-    assert trigsimp( (R2.r**2 - R2.x**2 - R2.y**2)(point_p) ) == 0
-
-    assert trigsimp(R2.e_r(R2.x**2+R2.y**2)(point_p).doit()) == 2*r0
-
-
-def test_intcurve_diffequ():
-    t = symbols('t')
-    start_point = R2_r.point([1, 0])
-    vector_field = -R2.y*R2.e_x + R2.x*R2.e_y
-    equations, init_cond = intcurve_diffequ(vector_field, t, start_point)
-    assert str(equations) == '[f_1(t) + Derivative(f_0(t), t), -f_0(t) + Derivative(f_1(t), t)]'
-    assert str(init_cond) == '[f_0(0) - 1, f_1(0)]'
-    equations, init_cond = intcurve_diffequ(vector_field, t, start_point, R2_p)
-    assert str(equations) == '[Derivative(f_0(t), t), Derivative(f_1(t), t) - 1]'
-    assert str(init_cond) == '[f_0(0) - 1, f_1(0)]'
